Dialogue UI Update
Changed font and dialogue box UI appearance.
Moved dialogue box over to allow for portraits later.
Also did an alternate dialogue system that allows for importing dialogue from separate text files; does not currently work 100% with the way things are coded so I left it unchecked for now, but might be interesting to incorporate in Dialogue 0.2.